requirements.txt
matplotlib==3.7.1
tkinter== 8.6.14
mysql-connector-python==9.2.0
